+++ domain/src/main/kotlin/com/kg/gettransfer/domain/model/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.kg.gettransfer.domain.model
 
-data class Coordinate(val transferId: Long? = 0, val lat: Float, val lon: Float)
+data class Coordinate(val transferId: Long? = 0, val lat: Double, val lon: Double)
